backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.9
pydantic==2.8.2
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.2
